requirements.txt
mss>=9.0.1
opencv-python>=4.8.0
numpy>=1.26.0
Pillow>=10.0.0
pyautogui>=0.9.54
pytesseract>=0.3.10

